fixup! IBX-7579:Richtext: Rows are added to ezurl_object_link on every save: Changed unit test into integration test
