+++ eZ/Publish/Core/FieldType/Tests/Url/Gateway/DoctrineStorageTest.php
@@ -194,29 +194,6 @@ class DoctrineStorageTest extends TestCase
         $this->assertEquals($expected, $result);
     }
 
-    /**
-     * @covers \eZ\Publish\Core\FieldType\Url\UrlStorage\Gateway\DoctrineStorage::getUrlsFromUrlLink
-     */
-    public function testGetUrlsFromUrlLink()
-    {
-        $gateway = $this->getStorageGateway();
-
-        $urlIds = [];
-        $urlIds[] = $gateway->insertUrl('https://ibexa.co/example1');
-        $urlIds[] = $gateway->insertUrl('https://ibexa.co/example2');
-        $urlIds[] = $gateway->insertUrl('https://ibexa.co/example3');
-
-        $gateway->linkUrl($urlIds[0], 10, 1);
-        $gateway->linkUrl($urlIds[1], 10, 1);
-        $gateway->linkUrl($urlIds[1], 12, 2);
-        $gateway->linkUrl($urlIds[2], 14, 1);
-
-        self::assertEquals(['https://ibexa.co/example1' => true, 'https://ibexa.co/example2' => true], $gateway->getUrlsFromUrlLink(10, 1), 'Did not get expected urlS for field 10');
-        self::assertEquals(['https://ibexa.co/example2' => true], $gateway->getUrlsFromUrlLink(12, 2), 'Did not get expected url for field 12');
-        self::assertEquals(['https://ibexa.co/example3' => true], $gateway->getUrlsFromUrlLink(14, 1), 'Did not get expected url for field 14');
-        self::assertEquals([], $gateway->getUrlsFromUrlLink(15, 1), 'Expected no urls for field 15');
-    }
-
     protected function getStorageGateway(): Gateway
     {
         if (!isset($this->storageGateway)) {
